🚚 [Rename]: place DTO 파일 폴더로 정리
