+++ src/place-review/review.service.ts
@@ -3,8 +3,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Builder } from 'builder-pattern';
 import { S3Service } from 'src/aws/s3/s3.service';
 import { BaseException } from 'src/global/base/base-exception';
-import { CreateReviewRequest } from 'src/place/dto/create-review-request.dto';
-import { CreateReviewResponse } from 'src/place/dto/create-review-response.dto';
+import { CreateReviewRequest } from 'src/place/dto/request/create-review-request.dto';
+import { CreateReviewResponse } from 'src/place/dto/response/create-review-response.dto';
 import { Place } from 'src/place/entity/place.entity';
 import { PlaceResponseCode } from 'src/place/exception/place-response-code';
 import { User } from 'src/user/entity/user.entity';
