+++ src/place/dto/response/place-detail-response.dto.ts
