+++ src/place/dto/response/place-list-response.dto.ts
