+++ src/place/place.service.ts
@@ -7,9 +7,9 @@ import { BaseException } from 'src/global/base/base-exception';
 import { GlobalResponseCode } from 'src/global/base/global-respose-code';
 import { ReviewImage } from 'src/place-review/entity/review-image.entity';
 import { Repository } from 'typeorm';
-import { PlaceDetailResponse } from './dto/place-detail-response.dto';
-import { PlaceListResponse } from './dto/place-list-response.dto';
-import { PlaceNewsResponse } from './dto/place-news-response.dto';
+import { PlaceDetailResponse } from './dto/response/place-detail-response.dto';
+import { PlaceListResponse } from './dto/response/place-list-response.dto';
+import { PlaceNewsResponse } from './dto/response/place-news-response.dto';
 import { PlaceBookmark } from './entity/place-bookmark.entity';
 import { PlaceHour } from './entity/place-hour.entity';
 import { PlaceRepository } from './entity/place.repository';
